% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateStatistics.R
\name{calculate_summary}
\alias{calculate_summary}
\title{CalculateStatistics}
\usage{
calculate_summary(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A list with mean, median, Q1, Q3, and IQR
}
\description{
Calculates a summary of the dataset
}
