% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectOutliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{DetectOutliers}
\usage{
detect_outliers(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A logical vector indicating outliers
}
\description{
Detects outliers
}
